+++ Makefile
@@ -36,7 +36,7 @@ test:
 
 test-tparse:
 	go test -race -count=1 ./internal/... -json -cover | go run main.go -trimpath=auto -sort=elapsed
-	go test -race -count=1 ./tests/... -json -cover -coverpkg=./parse | go run main.go -trimpath=github.com/mfridman/tparse/ -sort=elapsed
+	go test -race -count=1 ./tests/... -json -cover -coverpkg=./internal/parse | go run main.go -trimpath=github.com/mfridman/tparse/ -sort=elapsed
 
 # dogfooding :)
 test-tparse-full:
